pyproject.toml
[project]
name = "oai-supervisor"
version = "0.1.0"
description = ""
authors = [
    {name = "Rick Ross",email = "rick.ross@temporal.io"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio (>=3.4.3,<4.0.0)",
    "uuid (>=1.30,<2.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "aiohttp (>=3.12.14,<4.0.0)",
    "openai-agents == 0.3.2",
    "temporalio (>=1.18.0,<2.0.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "ibm-db-sa (>=0.4.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
